# bookstore-backend/openapi.yaml
openapi: 3.0.3
info:
  title: BookStore E-Commerce API
  description: |
    REST API for the BookStore E-Commerce capstone project.
    Provides endpoints for authentication, books catalogue, shopping cart,
    orders, payments, addresses, and gift points.
  version: 1.0.0
  contact:
    name: BookStore Support
    email: support@bookstore.com
servers:
  - url: http://localhost:5000/api
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List user orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders list
    post:
      tags: [Orders]
      summary: Place a new order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [addressId, paymentMethod]
              properties:
                addressId:
                  type: string
                  format: uuid
                paymentMethod:
                  type: string
                  enum: [card, upi, net_banking, cod]
                giftPointsToRedeem:
                  type: integer
                  minimum: 0
                  default: 0
      responses:
        '201':
          description: Order created
        '400':
          description: Empty cart or insufficient stock
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
        '404':
          description: Order not found
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order (within 48 hours)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled
        '400':
          description: Cannot cancel (already delivered, shipped, or >48 hours)
        '404':
          description: Order not found

  # ── ADDRESSES ──────────────────────────────────
...